+++ klum-ast/src/test/groovy/com/blackbuild/groovy/configdsl/transform/SpecialCasesSpec.groovy
@@ -60,7 +60,7 @@ class DynamicModel {
 }
 '''
         when:
-        instance = clazz.create().bla()
+        instance = clazz.Create.With().bla()
 
         then:
         clazz.count == 1